requirements.txt
fastapi
python-multipart
uvicorn
openai
pydantic
python-dotenv